Refactor authentication context and update components to use new context structure
