+++ src/auth/authContext.js
@@ -1,0 +1,5 @@
+import { createContext, useContext } from 'react';
+
+export const AuthContext = createContext();
+
+export const useAuth = () => useContext(AuthContext);
