--- src/hooks/useAuth.js
@@ -1,4 +1,0 @@
-import { useContext } from 'react';
-import AuthContext from '../auth/AuthProvider';
-
-export const useAuth = () => useContext(AuthContext);
